# test 1024^3 matrix mulmat performance