Summary:
Test the approach to improve 1024^3 matmul performance. On cooper lake, it can achieve 370GFlops/s, much better than 270Gflops/s before, close to 390 GFlops/s for MKL-DNN.
Probably need to add some prefetching to get 390 GFlops/s.

However, it significantly slows down the performance of other matrix shapes.
Need to figure out an approach for all matrix shapes.
This diff is for insight only now.

Probably we need to perform tuning inside getOrCreate function instead of inside
genComputeBlock function.

After removing time breakdown measurement, the performance difference for relatively small shapes becomes much smaller, roughly within 10%. For relatively larger shapes, performance keeps much higher.

Reviewed By: jianyuh

Differential Revision: D20472356